working on survey component
