+++ client/package.json
@@ -14,6 +14,7 @@
         "react-google-charts": "^3.0.15",
         "react-router-dom": "^5.2.0",
         "react-scripts": "4.0.1",
+        "survey-react": "^1.8.20",
         "web-vitals": "^0.2.4"
     },
     "scripts": {
