+++ client/public/homepage.html
@@ -67,5 +67,5 @@ body{
       Made with ❤️️ in Minnesota 
   </h3>
   <p style="text-align: center;"> 
-    By Jorge Calderon, Lierin Hanuman, James Jorissen,  Isabell Danell and Anusha Dhamera
+    By Jorge Calderon, Lierin Hanuman, James Jorissen,  Isabell Danell, Foos Mahamud and Anusha Dhamera
   </p>
