+++ client/src/App.js
@@ -3,6 +3,8 @@ import { BrowserRouter as Router, Route } from "react-router-dom";
 import NavBar from './components/NavBar';
 import HomePage from './components/HomePage';
 import Dashboard from './components/Dashboard';
+// import Signup from "./components/SignUpForm";
+import PHQ9 from "./components/PHQ9"
 
 function App() {
     return (
@@ -11,6 +13,8 @@ function App() {
                 <NavBar />
                 <Route exact path="/" component={HomePage} />
                 <Route exact path="/dashboard" component={Dashboard} />
+                {/* <Route exact path="/signup" component={Signup} /> */}
+                <Route exact path="/survey" component={PHQ9} />
             </div>
         </Router>
     );
